# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ControlRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Expected Revision",
            "type": "string"
          },
          "operation_id": {
            "pattern": "^[0-9a-f]{32}$",
            "title": "Operation Id",
            "type": "string"
          }
        },
        "required": [
          "operation_id",
          "expected_revision"
        ],
        "title": "ControlRequest",
        "type": "object"
      },
      "GitHubItem": {
        "additionalProperties": {
          "$ref": "#/components/schemas/JsonValue"
        },
        "properties": {
          "author": {
            "title": "Author",
            "type": "string"
          },
          "checks": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checks"
          },
          "head_sha": {
            "title": "Head Sha",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "observed_at": {
            "title": "Observed At",
            "type": "string"
          },
          "state": {
            "enum": [
              "open",
              "closed",
              "merged"
            ],
            "title": "State",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "number",
          "title",
          "state",
          "author",
          "url",
          "updated_at",
          "observed_at",
          "head_sha"
        ],
        "title": "GitHubItem",
        "type": "object"
      },
      "GitHubView": {
        "additionalProperties": {
          "$ref": "#/components/schemas/JsonValue"
        },
        "properties": {
          "coverage": {
            "title": "Coverage",
            "type": "string"
          },
          "has_more_remote": {
            "title": "Has More Remote",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/GitHubItem"
            },
            "title": "Items",
            "type": "array"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "never_synced": {
            "title": "Never Synced",
            "type": "boolean"
          },
          "next_cursor": {
            "title": "Next Cursor",
            "type": "string"
          },
          "observed_count": {
            "title": "Observed Count",
            "type": "integer"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "repository": {
            "title": "Repository",
            "type": "string"
          },
          "snapshot": {
            "title": "Snapshot",
            "type": "string"
          },
          "sources": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Sources",
            "type": "array"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "required": [
          "project_id",
          "repository",
          "kind",
          "items",
          "sources",
          "snapshot",
          "next_cursor",
          "has_more_remote",
          "total",
          "observed_count",
          "never_synced",
          "coverage"
        ],
        "title": "GitHubView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonValue": {},
      "Meta": {
        "properties": {
          "api_version": {
            "const": "1",
            "default": "1",
            "title": "Api Version",
            "type": "string"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "generated_at"
        ],
        "title": "Meta",
        "type": "object"
      },
      "NativeOperation": {
        "additionalProperties": {
          "$ref": "#/components/schemas/JsonValue"
        },
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "can_cancel": {
            "const": false,
            "default": false,
            "title": "Can Cancel",
            "type": "boolean"
          },
          "can_retry": {
            "const": false,
            "default": false,
            "title": "Can Retry",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "issue_number": {
            "title": "Issue Number",
            "type": "integer"
          },
          "last_error_code": {
            "title": "Last Error Code",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "pull_number": {
            "title": "Pull Number",
            "type": "integer"
          },
          "repository": {
            "title": "Repository",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "trace_command": {
            "title": "Trace Command",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "repository",
          "action",
          "state",
          "run_id",
          "issue_number",
          "pull_number",
          "updated_at",
          "last_error_code",
          "trace_command"
        ],
        "title": "NativeOperation",
        "type": "object"
      },
      "Operation": {
        "additionalProperties": {
          "$ref": "#/components/schemas/JsonValue"
        },
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "attempts": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Attempts",
            "type": "array"
          },
          "available_at": {
            "title": "Available At",
            "type": "string"
          },
          "binding_id": {
            "default": "",
            "title": "Binding Id",
            "type": "string"
          },
          "can_cancel": {
            "title": "Can Cancel",
            "type": "boolean"
          },
          "can_retry": {
            "title": "Can Retry",
            "type": "boolean"
          },
          "cancel_requested": {
            "default": false,
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "import_id": {
            "default": "",
            "title": "Import Id",
            "type": "string"
          },
          "last_error_code": {
            "title": "Last Error Code",
            "type": "string"
          },
          "max_attempts": {
            "title": "Max Attempts",
            "type": "integer"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "repository": {
            "title": "Repository",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "run_id": {
            "default": "",
            "title": "Run Id",
            "type": "string"
          },
          "stages": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Stages",
            "type": "array"
          },
          "stages_omitted": {
            "default": 0,
            "title": "Stages Omitted",
            "type": "integer"
          },
          "state": {
            "enum": [
              "pending",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "State",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "action",
          "repository",
          "state",
          "revision",
          "attempt_count",
          "max_attempts",
          "created_at",
          "updated_at",
          "available_at",
          "last_error_code",
          "can_cancel",
          "can_retry",
          "stages",
          "attempts"
        ],
        "title": "Operation",
        "type": "object"
      },
      "OperationList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Operation"
            },
            "title": "Items",
            "type": "array"
          },
          "native_items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/NativeOperation"
            },
            "title": "Native Items",
            "type": "array"
          },
          "native_next_before": {
            "default": 0,
            "title": "Native Next Before",
            "type": "integer"
          },
          "next_before": {
            "title": "Next Before",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "next_before"
        ],
        "title": "OperationList",
        "type": "object"
      },
...
      "Response_GitHubView_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GitHubView"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Response[GitHubView]",
        "type": "object"
      },
...
      "Response_OperationList_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/OperationList"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Response[OperationList]",
        "type": "object"
      },
      "Response_Operation_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Operation"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Response[Operation]",
        "type": "object"
      },
...
      "SyncRequest": {
        "additionalProperties": false,
        "properties": {
          "project_id": {
            "pattern": "^[0-9a-f]{32}$",
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id"
        ],
        "title": "SyncRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "RepoSteward local workbench",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/commands/github/sync": {
      "post": {
        "operationId": "syncGitHub",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9._:-]{1,128}$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_Operation_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync"
      }
    },
    "/api/v1/commands/operations/cancel": {
      "post": {
        "operationId": "cancelOperation",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9._:-]{1,128}$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_Operation_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel"
      }
    },
    "/api/v1/commands/operations/retry": {
      "post": {
        "operationId": "retryOperation",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9._:-]{1,128}$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_Operation_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry"
      }
    },
...
    "/api/v1/github": {
      "get": {
        "operationId": "github",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "pulls",
              "enum": [
                "pulls",
                "issues",
                "activity"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 100,
              "title": "Cursor",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "number",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 999999999,
              "minimum": 0,
              "title": "Number",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_GitHubView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Github"
      }
    },
...
    "/api/v1/operation": {
      "get": {
        "operationId": "operation",
        "parameters": [
          {
            "in": "query",
            "name": "operation_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Operation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_Operation_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Operation"
      }
    },
    "/api/v1/operations": {
      "get": {
        "operationId": "operations",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "default": "",
              "pattern": "^(?:[0-9a-f]{32})?$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Before",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_OperationList_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Operations"
      }
    },
...
  }
}